+++ src/redis/cluster/index.ts
@@ -1,0 +1,9 @@
+import { disallowDefaultNetwork } from './disallowDefaultNetwork'
+import { requireAuthorizationMode } from './requireAuthorizationMode'
+import { requireTransitEncryptionMode } from './requireTransitEncryptionMode'
+
+export const clusterPolicies = [
+	disallowDefaultNetwork,
+	requireAuthorizationMode,
+	requireTransitEncryptionMode
+]